@@ -1,22 +1,22 @@
 {
   "name": "lilplaytime/miniblog",
   "description": "Blog Web application. Backend for headless CMS, front end in Create-React-App",
-  "version": "2.4.4",
+  "version": "2.19.2",
   "license": "proprietary",
   "require": {
-    "php": "7.3.*",
-    "slim/slim": "2.*",
+    "php": "8.*.*",
+    "slim/slim": "4.*",
     "gabordemooij/redbean": "dev-master",
-    "vlucas/phpdotenv": "^5.4"
+    "vlucas/phpdotenv": "^5.4",
+    "nyholm/psr7": "^1.5",
+    "nyholm/psr7-server": "^1.0"
   },
   "require-dev": {
-    "codeception/codeception": "^4.1",
-    "codeception/module-phpbrowser": "^1.0.0",
-    "codeception/module-asserts": "^1.0.0",
-    "codeception/module-rest": "^1.2",
-    "squizlabs/php_codesniffer": "3.*",
-    "phpmd/phpmd": "@stable",
-    "zircote/swagger-php": "^3.0"
+    "codeception/codeception": "^5",
+    "codeception/module-phpbrowser": "^3.0",
+    "codeception/module-asserts": "^3.0",
+    "squizlabs/php_codesniffer": "^3.7",
+    "phpmd/phpmd": "^2.13"
   },
   "autoload": {
     "classmap": [
@@ -39,11 +39,11 @@
     "test": "vendor/bin/codecept run --steps",
     "test:acceptance": "vendor/bin/codecept run acceptance",
     "test:api": "vendor/bin/codecept run api",
-    "cs": "vendor/bin/phpcs -d standard=PSR2 backend/",
-    "cbf": "vendor/bin/phpcbf -d standard=PSR2 backend/",
-    "md": "vendor/bin/phpmd backend/ text codesize",
+    "cs": "vendor/bin/phpcs --standard=PSR2 --warning-severity=8 --ignore=*/vendor/*,*/build/* ./",
+    "csfix": "vendor/bin/phpcbf --standard=PSR2 --ignore=*/vendor/*,*/build/* ./",
+    "md": "vendor/bin/phpmd ./ text codesize,cleancode --exclude vendor/,*_generated*,build/*",
     "doc": "composer run doc:yaml && composer run doc:html",
-    "doc:yaml": "./vendor/bin/openapi --output doc/openapi.yaml backend/",
+    "doc:yaml": "./vendor/bin/openapi --output doc/openapi.yaml ./",
     "doc:html": "python yaml2html.py < doc/openapi.yaml > doc/doc.html"
   }
 }
