@@ -1,6 +1,20 @@
 <?php
-defined('ABSPATH') OR exit('No direct script access allowed');
+/**
+ * Model for ORM Entry
+ *
+ * PHP version 8
+ *
+ * @category PHP
+ * @package  Miniblog
+ * @author   Raymond Lim <rayjlim@yahoo.com>
+ * @license  MIT License
+ * @link     https://github.com/rayjlim/miniblog/
+ */
+namespace models;
+
+defined('ABSPATH') or exit('No direct script access allowed');
 use \Lpt\DevHelp;
+
 /**
  * @OA\Schema(
  *   schema="SmsEntrie",
@@ -10,40 +24,44 @@ use \Lpt\DevHelp;
 class SmsEntrie
 {
      /**
-     * The entry id
-     * @var integer
-     * @OA\Property()
-     */
-    public $id;
+      * The entry id
+      *
+      * @var           integer
+      * @OA\Property()
+      */
+    public int $id;
          /**
-     * The entry content
-     * @var string
-     * @OA\Property()
-     */
-    public $content;
+          * The entry content
+          *
+          * @var           string
+          * @OA\Property()
+          */
+    public string $content;
          /**
-     * The entry date
-     * @var date
-     * @OA\Property()
-     */
+          * The entry date
+          *
+          * @var           date
+          * @OA\Property()
+          */
     public $date;
          /**
-     * The entry creator id
-     * @var integer
-     * @OA\Property()
-     */
-    public $userId;
-    
+          * The entry creator id
+          *
+          * @var           integer
+          * @OA\Property()
+          */
+    public int $userId;
+
     public function __construct($date = '', $content = '')
     {
         $this->date = $date;
         $this->content = $content;
     }
-    
+
     public static function sanitizeContent($content)
     {
-        $order = array("\r\n", "\n", "\r");
-        
+        $order = ["\r\n", "\n", "\r"];
+
         // Processes \r\n's first so they aren't converted twice.
         $replace = '<br />';
         return stripslashes(str_replace($order, $replace, $content));
