@@ -1,15 +1,18 @@
 <?php
 namespace Lpt;
-defined('ABSPATH') OR exit('No direct script access allowed');
+
+defined('ABSPATH') or exit('No direct script access allowed');
+
+use \dao\Resource;
+
 class Logger
 {
     public static function log($message)
     {
-            $iResource = new \Resource();
+            $iResource = new Resource();
             $date = $iResource->getDateTime();
             $filename = LOGS_DIR.DIR_SEP.LOG_PREFIX."-" . $date->format(YEAR_MONTH_FORMAT).".txt";
             $fileData = $date->format(FULL_DATETIME_FORMAT) . "    " . $message."\n";
             $iResource->writeFile($filename, $fileData);
-
     }
 }
