@@ -1,10 +1,9 @@
 <?php
-function getValue($array, $key) {
-    if (!isset($array[$key])) {
-        return false;
-    }
-    return $array[$key];
+function getValue($array, $key)
+{
+    return $array[$key] ?? false;
 }
-function getExportValues($item) {
+function getExportValues($item)
+{
     return $item->export();
 }
