@@ -1,38 +1,50 @@
 <?php
-defined('ABSPATH') OR exit('No direct script access allowed');
+namespace dao;
 
+defined('ABSPATH') or exit('No direct script access allowed');
+
+use \models\SmsEntrie;
+use \models\ListParams;
 interface SmsEntriesDAO
 {
-
     /**
      * Get Domain object by primry key
      *
-     * @param String $id primary key
-     * @Return SmsEntries
+     * @param  string $id primary key
+     * @return array[SmsEntries]
      */
-    public function load($id);
-    
+    public function load(int $id): array;
+
     /**
      * Delete record from table
-     * @param smsEntrie primary key
+     *
+     * @param int Id of smsEntrie
      */
-    public function delete($id);
-    
+    public function delete(int $id): int;
+
     /**
      * Insert record to table
      *
-     * @param SmsEntries smsEntrie
+     * @param SmsEntrie Data for insertion
+     * @return int Id of new entry
      */
-    public function insert($smsEntrie);
-    
+    public function insert(SmsEntrie $smsEntrie): int;
+
     /**
      * Update record in table
      *
-     * @param SmsEntries smsEntrie
+     * @param array smsEntrie
      */
-    public function update($smsEntrie);
+    public function update(array $smsEntrie): void;
 
-    public function queryGraphData($userId, $graphParams);
-    public function queryBlogList($userId, $listParams);
-    public function queryLastTagEntry($userId, $label);
+    /**
+     * List Journal Entries
+     *
+     * @param  listParams search options
+     */
+    public function list(ListParams $listParams): array;
+
+    public function getSameDayEntries(object $date): array;
+
+    public function getYearMonths(): array;
 }
