@@ -1,10 +1,12 @@
 <?php
-defined('ABSPATH') OR exit('No direct script access allowed');
+namespace helpers;
+
+defined('ABSPATH') or exit('No direct script access allowed');
+use \models\SmsEntrie;
 
 interface ContentHelperInterface
 {
     public function __construct($_iDao, $_iResource);
 
     public function processEntry(SmsEntrie $smsEntry);
-
 }
